Rename some files and add dataclass
