+++ requirements.txt
@@ -6,6 +6,7 @@ bs4==0.0.1
 certifi==2022.12.7
 charset-normalizer==3.0.1
 click==8.1.3
+dataclasses==0.6
 exceptiongroup==1.1.0
 Flask==2.2.3
 h11==0.14.0
